Add coverage tests and refactor publish corner flow
