+++ frontend/src/components/publish/PublishCornerModal/components/LocationStep.types.ts
@@ -1,0 +1,16 @@
+import { TFunction } from 'i18next'
+
+import { PublishCornerFormState } from '../PublishCornerModal.types'
+
+export type LocationStepErrors = Partial<
+  Record<'addressSearch' | 'street' | 'number' | 'photo' | 'consent', string>
+>
+
+export type LocationStepProps = {
+  t: TFunction
+  state: PublishCornerFormState
+  errors: LocationStepErrors
+  onChange: (update: Partial<PublishCornerFormState>) => void
+  onPhotoSelect: (files: FileList | null) => void
+  onRemovePhoto: () => void
+}
